key creation changes
triplestore functions now take in Blobs instead of std::strings
    removed 1 alloc + memcpy per S and P

key format has changed to:
    S + P + '\xff' + S.len + P.len + S.type + P.type
This allows for sorting with the actual S + P, without data in between. Additionally, the '\xff' separator between the S + P and the rest of the key allows for comparision where S1 + P1 is a prefix of S2 + P2 to terminate, causing S1 + P1 to be sorted first.

Updated Blob constructor that takes in a string to also take in a type, for when the type is not a string

Updated tests